DEV-255 Refactor files for rebus address converter util
